+++ src/bindings/python/src/pyopenvino/graph/types/element_type.cpp
@@ -48,7 +48,6 @@ void regclass_graph_Type(py::module m) {
     type.attr("u32") = ov::element::u32;
     type.attr("u64") = ov::element::u64;
     type.attr("bf16") = ov::element::bf16;
-    type.attr("dynamic") = ov::element::dynamic;
 
     type.def("__hash__", &ov::element::Type::hash);
     type.def("__repr__", [](const ov::element::Type& self) {
