+++ src/bindings/python/tests/mock/mock_py_frontend/include/mock_py_frontend/frontend_wrappers.hpp
@@ -15,10 +15,6 @@
 #ifdef ENABLE_OV_TF_FRONTEND
 #    include "openvino/frontend/tensorflow/frontend.hpp"
 #endif
-
-#ifdef ENABLE_OV_PYTORCH_FRONTEND
-#    include "openvino/frontend/pytorch/frontend.hpp"
-#endif
 #include "visibility.hpp"
 
 #ifdef ENABLE_OV_ONNX_FRONTEND
