+++ src/frontends/common/src/manager.cpp
@@ -78,9 +78,8 @@ public:
         // Load all not loaded plugins/frontends
         std::lock_guard<std::mutex> guard(m_loading_mutex);
         for (auto& plugin_info : m_plugins) {
-            std::cerr << "Trying to load frontend: " << plugin_info.m_file_path << "\n";
             if (!plugin_info.load()) {
-                std::cerr << "FAILED\n";
+                std::cerr << "Frontend load failed: " << plugin_info.m_file_path << "\n";
                 continue;
             }
             names.push_back(plugin_info.get_creator().m_name);
