+++ src/frontends/pytorch/include/openvino/frontend/pytorch/node_context.hpp
@@ -1,3 +1,7 @@
+// Copyright (C) 2018-2022 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
 #pragma once
 
 #include <openvino/frontend/exception.hpp>
