+++ src/frontends/pytorch/src/exception.hpp
@@ -1,3 +1,6 @@
+// Copyright (C) 2018-2022 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
 
 #pragma once
