+++ src/frontends/pytorch/src/op_table.cpp
@@ -1,6 +1,7 @@
 // Copyright (C) 2018-2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
+
 #include "op_table.hpp"
 
 #include <openvino/opsets/opset8.hpp>
