+++ src/frontends/pytorch/src/pt_framework_node.hpp
@@ -1,3 +1,7 @@
+// Copyright (C) 2018-2022 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
 #include <openvino/op/util/framework_node.hpp>
 
 #include "utils.hpp"
