+++ src/frontends/pytorch/src/transforms.cpp
@@ -1,3 +1,7 @@
+// Copyright (C) 2018-2022 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
 #include "transforms.hpp"
 
 #include <openvino/frontend/pytorch/decoder.hpp>
